initial commit on gml parsing/compiling!
not ready for use yet - in fact, this part by itself isn't that useful
in general, unless you've ever thought "I wish that, instead of just
writing a lua program, I could code in an xml-based format with embedded
lua..." in which case you should seek therapy.

Ahem. Tangent aside...

test.lua loads gmltest.gml, which is a hand-written gml-version of the
gmltest.lua example program, and after parsing it and building a
practical structure from it that I'll be using in a big project later,
then compiles it into a lua program, which is written to output.lua.

output.lua is structurally a bit different, because of autogeneration,
but is functionally the same as gmltest.lua.